Creating demo screen
